+++ app/code/ThirdParty/BlogArticle/Model/Resolver/Post.php
@@ -51,6 +51,7 @@ class Post implements ResolverInterface
             'title' => $item->getTitle(),
             'url_key' => $item->getUrlKey(),
             'content' => $item->getContent(),
+            'excerpt' => $item->getExcerpt(),
             'featured_image' => $item->getFeaturedImage(),
             'meta_title' => $item->getMetaTitle(),
             'meta_description' => $item->getMetaDescription(),
@@ -59,6 +60,7 @@ class Post implements ResolverInterface
             'tag_ids' => $item->getTagIds() ?: [],
             'creation_time' => $item->getCreationTime(),
             'update_time' => $item->getUpdateTime(),
+            'published_at' => $item->getPublishedAt(),
             'model' => $item,
         ];
     }