+++ app/code/ThirdParty/BlogArticle/Model/Resolver/RelatedPosts.php
@@ -33,6 +33,7 @@ class RelatedPosts implements ResolverInterface
                 'title' => $item->getTitle(),
                 'url_key' => $item->getUrlKey(),
                 'content' => $item->getContent(),
+                'excerpt' => $item->getExcerpt(),
                 'featured_image' => $item->getFeaturedImage(),
                 'meta_title' => $item->getMetaTitle(),
                 'meta_description' => $item->getMetaDescription(),
@@ -41,6 +42,7 @@ class RelatedPosts implements ResolverInterface
                 'tag_ids' => $item->getTagIds() ?: [],
                 'creation_time' => $item->getCreationTime(),
                 'update_time' => $item->getUpdateTime(),
+                'published_at' => $item->getPublishedAt(),
             ];
         }
         return $mapped;